+++ src/ui/Features/Video/TransparentSubtitles/TransparentSubtitlesViewModel.cs
@@ -579,7 +579,7 @@ public partial class TransparentSubtitlesViewModel : ObservableObject
         if (long.TryParse(arr[1].Trim(), out var f))
         {
             _processedFrames = f;
-            ProgressValue = (double)_processedFrames * 100.0 / JobItems[_jobItemIndex].TotalFrames;
+            ProgressValue = _processedFrames * 100.0 / JobItems[_jobItemIndex].TotalFrames;
         }
     }
